+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Added
+
+- Casting and Alloying Calculators now display the layout of the stacks for each smelting process
+
 ## [0.3.0-beta.7] - 2025-12-04
 
 ### Fixed